[#46] Add Highcharts proof of concept (heatmap)

+++ app/src/main/resources/db/migration/V1__Structure_Client.sql
@@ -1,7 +1,7 @@
 CREATE TABLE orders (
  orderID         SERIAL,
- origin          SMALLINT,
- destination     SMALLINT,
+ origin          INT,
+ destination     INT,
  entryDate       DATE,
  departureDate   DATE
 );
